cleaned up search_criteria_dictionary_util - the function that creates the query text file from formDict
now this text file only picks up itms in formDict that start with sc_. I also cleaned up the next function that uses that text file to filter the records. search_criteria_dictionary_util is also moved to utils_general so the one function is used in both investigations and recalls. I started to make some other changes to the recalls search page so user can search on verified by.